+++ relay-server/src/envelope.rs
@@ -31,6 +31,7 @@
 //! ```
 
 use relay_base_schema::project::ProjectKey;
+use relay_profiling::ProfileType;
 use std::borrow::Borrow;
 use std::collections::BTreeMap;
 use std::fmt;
@@ -415,6 +416,9 @@ pub enum AttachmentType {
     /// This attachment is processed by Relay immediately and never forwarded or persisted.
     Breadcrumbs,
 
+    // A prosperodump crash report (binary data)
+    Prosperodump,
+
     /// This is a binary attachment present in Unreal 4 events containing event context information.
     ///
     /// This can be deserialized using the `symbolic` crate see
@@ -452,6 +456,7 @@ impl fmt::Display for AttachmentType {
             AttachmentType::Minidump => write!(f, "event.minidump"),
             AttachmentType::AppleCrashReport => write!(f, "event.applecrashreport"),
             AttachmentType::EventPayload => write!(f, "event.payload"),
+            AttachmentType::Prosperodump => write!(f, "playstation.prosperodump"),
             AttachmentType::Breadcrumbs => write!(f, "event.breadcrumbs"),
             AttachmentType::UnrealContext => write!(f, "unreal.context"),
             AttachmentType::UnrealLogs => write!(f, "unreal.logs"),
@@ -470,6 +475,7 @@ impl std::str::FromStr for AttachmentType {
             "event.minidump" => AttachmentType::Minidump,
             "event.applecrashreport" => AttachmentType::AppleCrashReport,
             "event.payload" => AttachmentType::EventPayload,
+            "playstation.prosperodump" => AttachmentType::Prosperodump,
             "event.breadcrumbs" => AttachmentType::Breadcrumbs,
             "event.view_hierarchy" => AttachmentType::ViewHierarchy,
             "unreal.context" => AttachmentType::UnrealContext,
@@ -591,6 +597,12 @@ pub struct ItemHeaders {
     #[serde(default, skip)]
     ingest_span_in_eap: bool,
 
+    /// Tracks whether the item is a backend or ui profile chunk.
+    ///
+    /// NOTE: This is internal-only and not exposed into the Envelope.
+    #[serde(default, skip)]
+    profile_type: Option<ProfileType>,
+
     /// Other attributes for forward compatibility.
     #[serde(flatten)]
     other: BTreeMap<String, Value>,
@@ -668,6 +680,7 @@ impl Item {
                 sampled: true,
                 fully_normalized: false,
                 ingest_span_in_eap: false,
+                profile_type: None,
             },
             payload: Bytes::new(),
         }
@@ -720,7 +733,11 @@ impl Item {
             ItemType::Span | ItemType::OtelSpan => smallvec![(DataCategory::Span, 1)],
             // NOTE: semantically wrong, but too expensive to parse.
             ItemType::OtelTracesData => smallvec![(DataCategory::Span, 1)],
-            ItemType::ProfileChunk => smallvec![(DataCategory::ProfileChunk, 1)], // TODO: should be seconds?
+            ItemType::ProfileChunk => match self.headers.profile_type {
+                Some(ProfileType::Backend) => smallvec![(DataCategory::ProfileChunk, 1)],
+                Some(ProfileType::Ui) => smallvec![(DataCategory::ProfileChunkUi, 1)],
+                None => smallvec![],
+            },
             ItemType::Unknown(_) => smallvec![],
         }
     }
@@ -885,6 +902,16 @@ impl Item {
         self.headers.ingest_span_in_eap = ingest_span_in_eap;
     }
 
+    /// Returns the associated profile type of a profile chunk.
+    pub fn profile_type(&self) -> Option<ProfileType> {
+        self.headers.profile_type
+    }
+
+    /// Set the profile type of the profile chunk.
+    pub fn set_profile_type(&mut self, profile_type: ProfileType) {
+        self.headers.profile_type = Some(profile_type);
+    }
+
     /// Gets the `sampled` flag.
     pub fn sampled(&self) -> bool {
         self.headers.sampled
@@ -934,6 +961,7 @@ impl Item {
                     AttachmentType::AppleCrashReport
                     | AttachmentType::Minidump
                     | AttachmentType::EventPayload
+                    | AttachmentType::Prosperodump
                     | AttachmentType::Breadcrumbs => true,
                     AttachmentType::Attachment
                     | AttachmentType::UnrealContext