+++ CHANGELOG.md
@@ -4,7 +4,17 @@
 
 **Features**:
 
+- Add experimental playstation endpoint. ([#4555](https://github.com/getsentry/relay/pull/4555))
+- Add naver.me / Yeti spider on the crawler filter list. ([#4602](https://github.com/getsentry/relay/pull/4602))
 - Add the item type that made the envelope too large to invalid outcomes. ([#4558](https://github.com/getsentry/relay/pull/4558))
+
+**Bug Fixes**:
+
+- Separates profiles into backend and ui profiles. ([#4595](https://github.com/getsentry/relay/pull/4595))
+
+**Internal**:
+
+- Add ui chunk profiling data category. ([#4593](https://github.com/getsentry/relay/pull/4593))
 - Switch global rate limiter to a service. ([#4581](https://github.com/getsentry/relay/pull/4581))
 
 ## 25.3.0